@@ -13,4 +13,4 @@ dev_dependencies:
     sdk: flutter
 
 environment:
-  sdk: '>=1.19.0 <2.0.0'
+  sdk: ">=2.12.0 <3.0.0"
